# openapi.yaml
openapi: 3.2.0
info:
  title: SDI - Servicio Distribuido de Inventario
  summary: API de gestion de stock para integracion entre modulos ERP.
  description: '

    API REST del modulo de Stock del sistema SDI.


    Permite administrar productos, consultar stock, crear reservas,

    confirmar ventas, liberar reservas y procesar eventos de compra.


    Esta API esta pensada para integracion entre sistemas internos:

    - Modulo de Ventas

    - Modulo de Compras

    - Modulo de Stock

    - Servicios administrativos internos


    La primera version no incluye seguridad formal. El consumo debe realizarse

    en ambiente controlado de red interna o infraestructura protegida.

    '
  termsOfService: https://docs.sdi.local/terminos-api
  contact:
    name: Equipo SDI - Sistema Distribuido de Inventario
    email: gxbeto@gmail.com
  license:
    name: Uso de la catedra Programacion de Aplicaiciones en Red - Profesor Alfonso
      Fernandez
  version: 0.1.0
servers:
- url: http://127.0.0.1:8000
  description: Ambiente local de desarrollo
- url: http://sdi-pruebas.example.com:6000
  description: Ambiente de pruebas / integracion
paths:
...
  /stock/movimientos:
    get:
      tags:
      - stock
      summary: Consultar historial de movimientos de stock
      description: Devuelve la trazabilidad de compras, reservas, ventas, liberaciones
        y ajustes. Permite filtrar por producto, documento de referencia y tipo de
        movimiento.
      operationId: listarMovimientosStock
      parameters:
      - name: producto_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Producto Id
          description: Debe ser un identificador entero positivo.
      - name: documento_ref
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Documento Ref
      - name: tipo_movimiento
        in: query
        required: false
        schema:
          type: string
          enum:
          - COMPRA
          - RESERVA
          - VENTA
          - LIBERACION_RESERVA
          - AJUSTE
          title: Tipo Movimiento
        description: Tipo de movimiento de stock usado como filtro de consulta.
      responses:
        '200':
          description: Lista de movimientos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovimientoOut'
                title: Response Listar Movimientos Stock Movimientos Get
              example:
              - movimiento_id: 50
                producto_id: 1
                reserva_id: 101
                tipo_movimiento: VENTA
                origen: VENTAS
                documento_ref: FV-001-000123
                comprobante: STK-001-000001-6
                cantidad: 2
                stock_anterior: 10
                stock_posterior: 8
                observacion: Confirmación de venta desde reserva
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
  /stock/{producto_id}:
    get:
      tags:
      - stock
      summary: Consultar stock actual de un producto
      description: Devuelve las cantidades total, reservada y disponible del producto
        indicado.
      operationId: obtenerStock
      parameters:
      - name: producto_id
        in: path
        required: true
        schema:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
      responses:
        '200':
          description: Stock actual.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockOut'
              example:
                producto_id: 1
                cantidad_total: 10
                cantidad_reservada: 2
                cantidad_disponible: 8
        '404':
          description: El producto solicitado no existe.
          content:
            application/json:
              example:
                estado: ERROR
                codigo: PRODUCTO_NO_EXISTE
                mensaje: El producto 1 no existe.
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
  /stock/reservas:
    post:
      tags:
      - stock
      summary: Crear reserva de stock para uno o mas productos
      description: 'Valida productos activos y stock disponible, bloquea las filas
        de stock involucradas, incrementa la cantidad reservada y registra movimientos
        de tipo RESERVA.


        La operación debe tratar reintentos equivalentes de forma idempotente. Clave
        idempotente: documento_ref.'
      operationId: crearReserva
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservaCreate'
        required: true
      responses:
        '201':
          description: Reserva creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservaCreateOut'
              example:
                estado: RESERVADO
                documento_ref: FV-001-000123
                comprobante: STK-001-000001-6
                reservas:
                - reserva_id: 101
                  producto_id: 1
                  cantidad_reservada: 2
                  cantidad_disponible: 8
        '400':
          description: El producto existe, pero no puede operar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                estado: ERROR
                codigo: PRODUCTO_INACTIVO
                mensaje: El producto 1 esta inactivo.
        '404':
          description: El producto solicitado no existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                estado: ERROR
                codigo: PRODUCTO_NO_EXISTE
                mensaje: El producto 1 no existe.
        '409':
          description: Conflicto funcional, replay no coincidente, operación ya procesada
            o conflicto de concurrencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                stockInsuficiente:
                  summary: Stock insuficiente
                  value:
                    estado: ERROR
                    codigo: STOCK_INSUFICIENTE
                    mensaje: 'Stock insuficiente para el producto 1: solicitado 2,
                      disponible 1.'
                documentoYaTieneReserva:
                  summary: Documento ya reservado
                  value:
                    estado: ERROR
                    codigo: DOCUMENTO_YA_TIENE_RESERVA
                    mensaje: El documento FV-001-000123 ya posee una reserva registrada.
                replayNoCoincide:
                  summary: Replay con contenido diferente
                  value:
                    estado: ERROR
                    codigo: REPLAY_NO_COINCIDE
                    mensaje: El documento_ref ya existe, pero los ítems recibidos
                      no coinciden con la reserva original.
                conflictoConcurrencia:
                  summary: Conflicto de concurrencia
                  value:
                    estado: ERROR
                    codigo: CONFLICTO_CONCURRENCIA
                    mensaje: La reserva no pudo completarse porque el stock fue modificado
                      por otra operación concurrente.
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
      x-idempotency:
        enabled: true
        keyField: documento_ref
        scope: Documento externo de venta o pedido que solicita la reserva.
        replayBehavior: Si llega el mismo documento_ref con los mismos ítems, no se
          duplica la reserva; se devuelve el estado ya registrado.
        payloadConflictBehavior: Si llega el mismo documento_ref con ítems diferentes,
          se rechaza con 409 y código REPLAY_NO_COINCIDE.
        sameRequestPolicy: No debe duplicar movimientos, reservas ni descuentos ante
          reintentos equivalentes.
        differentPayloadPolicy: Debe responder 409 si la clave idempotente ya existe
          con contenido diferente.
      x-concurrency:
        strategy: pessimistic-locking
        conflictStatus: 409
        conflictCode: CONFLICTO_CONCURRENCIA
        description: Las operaciones que afectan stock deben ejecutarse en transacción
          y proteger el saldo/reserva contra modificaciones concurrentes.
  /stock/reservas/{reserva_id}/confirmar:
    post:
      tags:
      - stock
      summary: Confirmar reserva y descontar stock definitivo
      description: 'Confirma una reserva pendiente, descuenta cantidad_total y cantidad_reservada,
        registra movimiento de VENTA y opcionalmente precio de venta.


        Operación de uso interno. Solo debe ser invocada por módulos internos del
        ecosistema SDI, principalmente Ventas o procesos administrativos autorizados.
        No forma parte del contrato público para integradores externos.


        La operación debe tratar reintentos equivalentes de forma idempotente. Clave
        idempotente: reserva_id.'
      operationId: confirmarReserva
      parameters:
      - name: reserva_id
        in: path
        required: true
        schema:
          type: integer
          title: Reserva Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmarReservaIn'
      responses:
        '200':
          description: Reserva confirmada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmarReservaOut'
              example:
                estado: CONFIRMADA
                reserva_id: 101
                producto_id: 1
                documento_ref: FV-001-000123
                comprobante: STK-001-000002-4
                cantidad_descontada: 2
                stock_total_actual: 8
                cantidad_reservada_actual: 0
                cantidad_disponible_actual: 8
        '404':
          description: La reserva solicitada no existe.
          content:
            application/json:
              example:
                estado: ERROR
                codigo: RESERVA_NO_EXISTE
                mensaje: La reserva 101 no existe.
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto funcional, replay no coincidente, operación ya procesada
            o conflicto de concurrencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                operacionYaProcesada:
                  summary: Confirmación repetida
                  value:
                    estado: ERROR
                    codigo: OPERACION_YA_PROCESADA
                    mensaje: La reserva 101 ya fue confirmada previamente.
                reservaYaLiberada:
                  summary: Reserva liberada previamente
                  value:
                    estado: ERROR
                    codigo: RESERVA_YA_LIBERADA
                    mensaje: La reserva 101 ya fue liberada y no puede confirmarse.
                conflictoConcurrencia:
                  summary: Conflicto de concurrencia
                  value:
                    estado: ERROR
                    codigo: CONFLICTO_CONCURRENCIA
                    mensaje: La reserva fue modificada por otra operación concurrente.
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
      x-internal: true
      x-idempotency:
        enabled: true
        keyField: reserva_id
        scope: Confirmación interna de una reserva específica.
        replayBehavior: Si la reserva ya está CONFIRMADA, la repetición de la confirmación
          se trata como idempotente y devuelve el estado confirmado sin descontar
          stock nuevamente.
        payloadConflictBehavior: Si la reserva está LIBERADA o no corresponde a la
          operación, se rechaza con 409.
        sameRequestPolicy: No debe duplicar movimientos, reservas ni descuentos ante
          reintentos equivalentes.
        differentPayloadPolicy: Debe responder 409 si la clave idempotente ya existe
          con contenido diferente.
      x-concurrency:
        strategy: pessimistic-locking
        conflictStatus: 409
        conflictCode: CONFLICTO_CONCURRENCIA
        description: Las operaciones que afectan stock deben ejecutarse en transacción
          y proteger el saldo/reserva contra modificaciones concurrentes.
  /stock/reservas/{reserva_id}/liberar:
    post:
      tags:
      - stock
      summary: Liberar reserva pendiente y devolver disponibilidad
      description: 'Libera una reserva pendiente cuando la venta no se concreta, reduce
        cantidad_reservada y registra movimiento de LIBERACION_RESERVA.


        Operación de uso interno. Solo debe ser invocada por módulos internos del
        ecosistema SDI, principalmente Ventas o procesos administrativos autorizados.
        No forma parte del contrato público para integradores externos.


        La operación debe tratar reintentos equivalentes de forma idempotente. Clave
        idempotente: reserva_id.'
      operationId: liberarReserva
      parameters:
      - name: reserva_id
        in: path
        required: true
        schema:
          type: integer
          title: Reserva Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LiberarReservaIn'
      responses:
        '200':
          description: Reserva liberada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiberarReservaOut'
              example:
                estado: LIBERADA
                reserva_id: 101
                comprobante: STK-001-000003-2
                cantidad_liberada: 2
        '404':
          description: La reserva solicitada no existe.
          content:
            application/json:
              example:
                estado: ERROR
                codigo: RESERVA_NO_EXISTE
                mensaje: La reserva 101 no existe.
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto funcional, replay no coincidente, operación ya procesada
            o conflicto de concurrencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                operacionYaProcesada:
                  summary: Liberación repetida
                  value:
                    estado: ERROR
                    codigo: OPERACION_YA_PROCESADA
                    mensaje: La reserva 101 ya fue liberada previamente.
                reservaYaConfirmada:
                  summary: Reserva confirmada previamente
                  value:
                    estado: ERROR
                    codigo: RESERVA_YA_CONFIRMADA
                    mensaje: La reserva 101 ya fue confirmada y no puede liberarse.
                conflictoConcurrencia:
                  summary: Conflicto de concurrencia
                  value:
                    estado: ERROR
                    codigo: CONFLICTO_CONCURRENCIA
                    mensaje: La reserva fue modificada por otra operación concurrente.
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
      x-internal: true
      x-idempotency:
        enabled: true
        keyField: reserva_id
        scope: Liberación interna de una reserva específica.
        replayBehavior: Si la reserva ya está LIBERADA, la repetición de la liberación
          se trata como idempotente y devuelve el estado liberado sin alterar stock
          nuevamente.
        payloadConflictBehavior: Si la reserva está CONFIRMADA, se rechaza con 409.
        sameRequestPolicy: No debe duplicar movimientos, reservas ni descuentos ante
          reintentos equivalentes.
        differentPayloadPolicy: Debe responder 409 si la clave idempotente ya existe
          con contenido diferente.
      x-concurrency:
        strategy: pessimistic-locking
        conflictStatus: 409
        conflictCode: CONFLICTO_CONCURRENCIA
        description: Las operaciones que afectan stock deben ejecutarse en transacción
          y proteger el saldo/reserva contra modificaciones concurrentes.
...
components:
  schemas:
...
    ConfirmarReservaIn:
      properties:
        documento_ref:
          anyOf:
          - type: string
            maxLength: 14
            pattern: ^[A-Z]{2,3}-\d{3}-\d{6}$
            description: 'Documento externo recibido desde Ventas o Compras. Se usa
              para trazabilidad externa. Ejemplos: FV-001-000123, OC-001-000555.'
            examples:
            - FV-001-000123
          - type: 'null'
          title: Documento Ref
        precio_venta:
          anyOf:
          - type: number
            multipleOf: 0.01
            minimum: 0.0
            format: decimal
            example: 32000
          - type: 'null'
          title: Precio Venta
      type: object
      title: ConfirmarReservaIn
      example:
        documento_ref: FV-001-000123
        precio_venta: 45760
...
    ErrorResponse:
      properties:
        estado:
          type: string
          enum:
          - ERROR
          title: Estado
          description: Estado fijo para respuestas de error.
          default: ERROR
        codigo:
          type: string
          enum:
          - VALIDACION_ERROR
          - PRODUCTO_NO_EXISTE
          - PRODUCTO_INACTIVO
          - PRODUCTO_CODIGO_DUPLICADO
          - STOCK_NO_INICIALIZADO
          - STOCK_INSUFICIENTE
          - RESERVA_NO_EXISTE
          - RESERVA_NO_PENDIENTE
          - RESERVA_NO_COINCIDE
          - NO_AUTENTICADO
          - ACCESO_DENEGADO
          - ERROR_INTERNO
          - ERROR_NO_CONTROLADO
          - MENSAJE_DUPLICADO
          - REPLAY_NO_COINCIDE
          - OPERACION_YA_PROCESADA
          - OPERACION_EN_PROCESO
          - CONFLICTO_CONCURRENCIA
          - DOCUMENTO_YA_PROCESADO
          - DOCUMENTO_YA_TIENE_RESERVA
          - RESERVA_YA_CONFIRMADA
          - RESERVA_YA_LIBERADA
          title: Codigo
          description: Código técnico de error devuelto por la API.
        mensaje:
          type: string
          title: Mensaje
          examples:
          - Descripcion clara del error.
      type: object
      required:
      - codigo
      - mensaje
      title: ErrorResponse
      example:
        codigo: CODIGO_ERROR
        estado: ERROR
        mensaje: Descripcion clara del error.
...
    LiberarReservaIn:
      properties:
        motivo_liberacion:
          type: string
          maxLength: 200
          minLength: 1
          title: Motivo Liberacion
          examples:
          - Cancelacion de venta
      type: object
      required:
      - motivo_liberacion
      title: LiberarReservaIn
      example:
        motivo_liberacion: Cancelacion de venta
...
    MovimientoOut:
      properties:
        movimiento_id:
          type: integer
          title: Movimiento Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        producto_id:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        reserva_id:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Reserva Id
          description: Debe ser un identificador entero positivo.
        tipo_movimiento:
          type: string
          enum:
          - COMPRA
          - RESERVA
          - VENTA
          - LIBERACION_RESERVA
          - AJUSTE
          title: Tipo Movimiento
          description: Tipo de operación que generó el movimiento de stock.
        origen:
          type: string
          enum:
          - COMPRAS
          - VENTAS
          - STOCK
          - SISTEMA
          title: Origen
          description: Módulo o sistema que originó el movimiento.
        documento_ref:
          anyOf:
          - type: string
            pattern: ^[A-Z]{2,3}-\d{3}-\d{6}$
            description: 'Documento externo recibido desde Ventas o Compras. Se usa
              para trazabilidad externa. Ejemplos: FV-001-000123, OC-001-000555.'
            examples:
            - FV-001-000123
          - type: 'null'
          title: Documento Ref
        comprobante:
          type: string
          pattern: ^STK-\d{3}-\d{6}-\d$
          title: Comprobante
          description: Comprobante interno generado por el módulo Stock para identificar
            la operación dentro del sistema. Incluye dígito verificador para validación
            de entrada.
          examples:
          - STK-001-000001-6
        cantidad:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Cantidad
          examples:
          - 10
          example: 10
        stock_anterior:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Stock Anterior
          examples:
          - 10
          example: 10
        stock_posterior:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Stock Posterior
          examples:
          - 10
          example: 10
        observacion:
          anyOf:
          - type: string
          - type: 'null'
          title: Observacion
      type: object
      required:
      - movimiento_id
      - producto_id
      - reserva_id
      - tipo_movimiento
      - origen
      - comprobante
      - cantidad
      - stock_anterior
      - stock_posterior
      - observacion
      title: MovimientoOut
      example:
        cantidad: 2
        comprobante: STK-001-000001-6
        documento_ref: FV-001-000123
        movimiento_id: 50
        observacion: Confirmacion de venta desde reserva
        origen: VENTAS
        producto_id: 1
        reserva_id: 101
        stock_anterior: 10
        stock_posterior: 8
        tipo_movimiento: VENTA
...
    ReservaCreate:
      properties:
        documento_ref:
          type: string
          maxLength: 14
          pattern: ^[A-Z]{2,3}-\d{3}-\d{6}$
          title: Documento Ref
          description: 'Documento externo recibido desde Ventas o Compras. Se usa
            para trazabilidad externa. Ejemplos: FV-001-000123, OC-001-000555.'
          examples:
          - FV-001-000123
        items:
          items:
            $ref: '#/components/schemas/ReservaItemIn'
          type: array
          minItems: 1
          title: Items
      type: object
      required:
      - documento_ref
      - items
      title: ReservaCreate
      example:
        documento_ref: FV-001-000123
        items:
        - cantidad: 2
          producto_id: 1
    ReservaCreateOut:
      properties:
        estado:
          type: string
          enum:
          - RESERVADO
          title: Estado
          description: Estado devuelto cuando la reserva fue creada.
          default: RESERVADO
        documento_ref:
          type: string
          pattern: ^[A-Z]{2,3}-\d{3}-\d{6}$
          title: Documento Ref
          description: 'Documento externo recibido desde Ventas o Compras. Se usa
            para trazabilidad externa. Ejemplos: FV-001-000123, OC-001-000555.'
          examples:
          - FV-001-000123
        comprobante:
          type: string
          pattern: ^STK-\d{3}-\d{6}-\d$
          title: Comprobante
          description: Comprobante interno generado por el módulo Stock para identificar
            la operación dentro del sistema. Incluye dígito verificador para validación
            de entrada.
          examples:
          - STK-001-000001-6
        reservas:
          items:
            $ref: '#/components/schemas/ReservaItemOut'
          type: array
          title: Reservas
      type: object
      required:
      - estado
      - documento_ref
      - comprobante
      - reservas
      title: ReservaCreateOut
      example:
        comprobante: STK-001-000001-6
        documento_ref: FV-001-000123
        estado: RESERVADO
        reservas:
        - cantidad_disponible: 8
          cantidad_reservada: 2
          producto_id: 1
          reserva_id: 101
    ReservaItemIn:
      properties:
        producto_id:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        cantidad:
          type: number
          multipleOf: 0.001
          exclusiveMinimum: 0.0
          format: decimal
          title: Cantidad
          examples:
          - 2
          example: 2
      type: object
      required:
      - producto_id
      - cantidad
      title: ReservaItemIn
      example:
        cantidad: 2
        producto_id: 1
...
    StockOut:
      properties:
        producto_id:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        cantidad_total:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Cantidad Total
          examples:
          - 10
          example: 10
        cantidad_reservada:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Cantidad Reservada
          examples:
          - 10
          example: 10
        cantidad_disponible:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Cantidad Disponible
          examples:
          - 10
          example: 10
      type: object
      required:
      - producto_id
      - cantidad_total
      - cantidad_reservada
      - cantidad_disponible
      title: StockOut
      example:
        cantidad_disponible: 8
        cantidad_reservada: 2
        cantidad_total: 10
        producto_id: 1
...
  responses:
    ValidationErrorResponse:
      description: El request no cumple el contrato OpenAPI o las validaciones declaradas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: VALIDACION_ERROR
            mensaje: 'campo: descripción clara de la validación incumplida.'
    UnauthorizedErrorResponse:
      description: No autenticado. Falta credencial válida o el mecanismo de autenticación
        rechazó la solicitud.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: NO_AUTENTICADO
            mensaje: No se pudo autenticar la solicitud.
    ForbiddenErrorResponse:
      description: Acceso denegado. El consumidor no tiene permiso para ejecutar la
        operación solicitada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: ACCESO_DENEGADO
            mensaje: No tiene permisos para ejecutar esta operación.
    InternalServerErrorResponse:
      description: Error interno controlado del servicio.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: ERROR_INTERNO
            mensaje: Ocurrió un error interno al procesar la operación.
    DefaultErrorResponse:
      description: Error no controlado o no clasificado por el contrato.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: ERROR_NO_CONTROLADO
            mensaje: Ocurrió un error no clasificado por el contrato.
...